Create first draft of app
Add Firebase as a client.
Add Prettier and configure with ESLint.
Add components necessary to render a list of stories.

This is limited to a read-only view of just the top stories from HN.
I link out to HN for the discussion and author pages.
Previews are captured via the /api/preview route using a nice little library
to stream just the headers, and then they should be cached by Next.