Switch to split-panel UI
I wanted to be able to see the list of stories and read the story and comments at the same time.
This refactors a lot of the UI elements and switches them to client-side.
To do this, we now have to fetch a lot of items from the /api/items/[itemID] and related endpoints.
We faciliate mobile with a horizontally scrolling article list and a vertical article view.
We also fetch comments on articles and allow for nested comments.
Other aspects of the UI have been tightened up and added to with Heroicons.