Remove api routes for items
We can (and should) use firebase from the client so that we aren't hammering our api.
Remove the api routes and rely on swr to hold a local cache for items instead of
double caching there and in the api.
Add some spinners so things look a little cleaner when its loading.